Update aes.c
I noticed that the inverse mix column function was very complex and wanted to simplify it while tailoring the new function towards the constraints of the project, which is to keep it as condensed as possible.  The benefits of the new Multiply function I have suggested are that it is significantly easier for a user to understand how the finite field multiplication works, it eliminates the need for the xtime() function, and that it operates under close to identical speeds as the original Multiply function.

From my observation, the Inverse Mix Columns function is the most complex function in AES and is not particularly well documented online.  A simpler function may 
be beneficial towards those who may learn more about encryption/decryption.